Use SubjectPage for toolbox subjects and multidisciplinary subjects
