+++ src/containers/SubjectPage/SubjectRouting.tsx
@@ -11,8 +11,6 @@ import SubjectPage from "./SubjectPage";
 import { useUrnIds } from "../../routeHelpers";
 import FilmFrontpage from "../FilmFrontpage/FilmFrontpage";
 import MultidisciplinarySubjectArticlePage from "../MultidisciplinarySubject/MultidisciplinarySubjectArticlePage";
-import MultidisciplinarySubjectPage from "../MultidisciplinarySubject/MultidisciplinarySubjectPage";
-import ToolboxSubjectPage from "../ToolboxSubject/ToolboxSubjectPage";
 
 const SubjectRouting = () => {
   const { topicList, subjectType } = useUrnIds();
@@ -23,9 +21,9 @@ const SubjectRouting = () => {
     if (topicList.length === 3) {
       return <MultidisciplinarySubjectArticlePage />;
     }
-    return <MultidisciplinarySubjectPage />;
+    return <SubjectPage />;
   } else if (subjectType === "toolbox") {
-    return <ToolboxSubjectPage />;
+    return <SubjectPage />;
   } else if (subjectType === "film" && topicList.length === 0) {
     return <FilmFrontpage />;
   } else if (subjectType === "film") {
